RAB-1425: Fix Backuper integration test
